feat: Refactor BYOM configuration to support single model setup and enhance onboarding flow with hybrid provider options
